Rename BIP 151 files
In preparation for BIP 150 code inclusion (circular dependency hell will be encountered otherwise), rename BIP 151 files to indicate inclusion of 150 & 151 code. No 150 code has been check in just yet.
