+++ cppForSwig/BIP150_151.cpp
